svm: fix callstack memory bug & @main lookup
If the user had not properly returned from all functions they had
called, some stack frames would not be free'd properly. This has
been fixed now.

There was a problem if your function name started with 'main' it
counted it as a main function and therefor jumped to it on startup.