svm: implement EXTLD functionality & TRACE
The EXTLD instruction loads a new module into the module list,
making the functions inside it availble for other modules to use.
If a local ext/ directory does not exist, it looks inside the Salt
Home ext/ directory in an attempt to find the new module. You
can find detailed explanations in how this works in the SVM header
files.

The TRACE instruction prints the current stack trace to standard out.